Split `callApi` and `testApi` serverless handlers
