+++ packages/airnode-node/src/handlers/test-api.test.ts
@@ -1,6 +1,6 @@
 import { Endpoint } from '@api3/airnode-ois';
 import { testApi } from './test-api';
-import * as worker from '../adapters/http/worker';
+import * as api from '../api';
 import * as fixtures from '../../test/fixtures';
 
 const ENDPOINT_ID = '0x13dea3311fe0d6b84f4daeab831befbc49e19e6494c41e9e065a09c3c68f43b6';
@@ -45,36 +45,23 @@ describe('testApi', () => {
   });
 
   it('calls the API with given parameters', async () => {
-    const spy = jest.spyOn(worker, 'spawnNewApiCall');
+    const spy = jest.spyOn(api, 'callApi');
     spy.mockResolvedValueOnce([[], { success: true, value: '1000', signature: 'not used' }]);
 
     const parameters = { _type: 'int256', _path: 'price', from: 'ETH' };
     const [err, res] = await testApi(fixtures.buildConfig(), ENDPOINT_ID, parameters);
 
+    const config = fixtures.buildConfig();
     const aggregatedApiCall = fixtures.buildAggregatedTestingGatewayApiCall({
       airnodeAddress: '0xA30CA71Ba54E83127214D3271aEA8F5D6bD4Dace',
       endpointId: ENDPOINT_ID,
       id: expect.any(String),
       parameters,
     });
-    const logOptions = {
-      format: 'plain',
-      level: 'DEBUG',
-      meta: {
-        requestId: expect.any(String),
-      },
-    };
-    const workerOptions = {
-      cloudProvider: {
-        type: 'local',
-      },
-      airnodeAddressShort: expect.any(String),
-      stage: 'test',
-    };
 
     expect(err).toBeNull();
     expect(res).toEqual({ value: '1000', signature: 'not used', success: true });
     expect(spy).toHaveBeenCalledTimes(1);
-    expect(spy).toHaveBeenCalledWith(aggregatedApiCall, logOptions, workerOptions);
+    expect(spy).toHaveBeenCalledWith({ config, aggregatedApiCall });
   });
 });
